Remove onmount master and label scripts, script startup
Script was used previously to sync wallpaper when I mounted my DROID.

However I don't do that any longer, and have wireless sync set up for
the things I do currently.

The master script was a lot of work to create, so I restored it to get
it in version control before deleting it.