maildirproc: save a copy of roster emails in .archive.not
Most of the time roster emails are the weekly roster email that I just
want processed, but sometimes they are custom roster searches.  It's
hard to distinguish, and the latter is rare, so just copy all roster
emails to .archive.not where I can fish them out if I want them, and
where they'll expire if I don't.