firefox xfce launcher: support different HOME dir
Existing launchers had paths for Exec and Icon that hard-coded
/home/edgewood, but these need to be usable for /home/edblac at work.
Use relative paths in both, relying on the firefox-profile script to be
in $PATH for Exec, and the icons being in $HOME/.icons for Icon.

The Icon Theme Specification at
https://specifications.freedesktop.org/icon-theme-spec/icon-theme-spec-latest.html
says that $HOME/.icons is the first search path if not using an absolute
path.