maildirproc: remove rc_data_files references
A previous version of maildirproc (perhaps a custom version I never sent
a patch for?) had a public list of config files that were all monitored
for changes, the way that the default.rc file usually is.  That allowed
other files to effectively become part of the default.rc in terms of
restarting maildirproc if they were changed.

I don't care to restore that functionality to the current version of
maildirproc (at this time), so need to remove the references the list
that no longer exists.