todo/helpers/normalize_date: support past/future bias
When normalizing the date, the incoming date text may not specify a
year, but there's one implicit in the context where the normalization
was called from: dates from 'do' are in the past, and 'postpone' dates
should be in the future.

Change normalize date so that the calling code can pass in that context,
and adjust the date accordingly.