profile: don't set LOCATE_PATH if no *.db files in ~/.cache/mlocate
Old code would set LOCATE_PATH to "$HOME/.cache/mlocate/*.db" if there
were no *.db files, causing errors.  The fix is to set the path in a way
that will be empty if ~/.cache/mlocate doesn't exist or has no *.db
files[1].

This change also unsets LOCATE_PATH if it's empty, since a set-but-empty
LOCATE_PATH causes locate to print duplicate messages for matches in the
system database.

[1] When the system was unencrypted, but my home directory was
(ecryptfs), I excluded my home directory from the system locate scans
and had a separate scan of just the home directory when it was
unecrypted, to avoid scanning the encrypted files in the home directory.
Now that the whole system is on an encrypted partition, there's no
chance of the system running and finding encrypted home directory files,
so that configuration is no longer enabled.  However, setting
LOCATE_PATH might be useful in the future, so I'm keeping this
low-impact check in place.