bashrc: add imv function
Add interactive mv function that takes a (probably tab-completed or
pasted) source path and prompts for the destination path using the
source path as the initial value.  This allows interactively making
changes to the path, handy for times when the target path is some
modification of the source.

The usual command line editing facilities (Ctrl-W, tab-completion) are
available when editing the prompt.